authentication and authorization middleware

+++ entity/src/black_list.rs
@@ -7,8 +7,8 @@ use serde::{Deserialize, Serialize};
 #[sea_orm(table_name = "black_list")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i64,
-    pub user_id: i64,
+    pub id: i32,
+    pub user_id: i32,
     #[sea_orm(column_type = "Text")]
     pub note_text: String,
 }
